From: Holger Macht <hmacht@suse.de>, Luca Tettamanti <kronos.it@gmail.com>
The current DirectFB implementation handles keyboard events in a separate
thread and this thread is frozen along with all other userspace during
image writing. Because of this, keyboard presses are blocked when splashy
is active which prevents aborting suspend or setting the reboot
mode. Patch changes the keyboard handling and reads all keyboard events
from /dev/input/by-path/*kbd instead of from stdin. This makes the
keyboard handling independent from any splash implementation.

Additionally, '#define KEY_FILE' is changed to '#define KEY_FILE_PATH' to
not conflict with the content of linux/input.h.

Further information and references are here:

http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=411061
https://bugzilla.novell.com/show_bug.cgi?id=293826