From: Alon Bar-Lev <alon.barlev@gmail.com>
Add fbsplash support for suspend/resume.

fbsplash has two modes: silent and verbose.
In order to change the caption of the silent splash I added
a new callback for the splash module: set_caption().

I modified the important messages so that it will be printed to the
verbose (stdout) as well to the silent.