It turns out that we can cut the times of saving compressed images by half, so that they are comparable with the times of loading these images.
Unfortunately, for this purpose it's necessary to rework the saving and loading
of the images to use some other type of metadata and to arrange things so that
blocks larger than PAGE_SIZE are compressed.

With this patch applied, we can save approximately twice as much image data
data as before over a time unit.  For example, on one of my test systems it
saves around 82 MB of image data per second and loads 85 MB/s.  With encryption
we can write about 40 MB/s and read about 48 MB/s on the same system.  For the
unpatched code on the same machine I got 36 MB/s, 84 MB/s, 23 MB/s, 45 MB/s,
respectively.

On another, slightly faster test system, I get write and read speeds above
100 MB/s with this patch applied.