Add configuration option "theads = " that makes s2disk/s2both use additional threads for saving the image, which may speed up the saving of compressed and encrypted images on multicore boxes.
The idea is that the image pages read from the kernel are first placed in a
work buffer, which is then compressed and its (compressed) contents are written
into one of four "write" buffers (those buffers are used in a round robin
manner).  Then, another thread encrypts the contents of the work buffers and
places the (encrypted) data into a "save" buffer that is read by the third
thread which finally writes the data to disk.

It turns out that with this patch applied one can speed up the saving of
encrypted images quite significantly.  For example, on my test box the speed
of "threaded" saving of an encrytpded image is of the order of 60 MB/s,
while for the same operation without threads the average speed is about
40 MB/s.