Fix "program swap-offset is using a deprecated SCSI ioctl"
"program swap-offset is using a deprecated SCSI ioctl"

<http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=425219>

The debian uswsusp package tries to run swap-offset even if the
swap is a partition and not a file.  Calling FIGETBSZ on a scsi disk
device provokes the alarming kernel warning above.  (Alarming
because it suggests that FIGETBSZ might somehow be interpreted as a
random SCSI ioctl).

Let's check explicitly whether we've been given a regular file.

Signed-off-by: Rafael J. Wysocki <rjw@sisk.pl>