s2ram: Do not apply quirks if KMS driver is in use
If there's a KMS graphics driver in use, s2ram shouldn't really do
anything apart from telling the kernel to suspend.  Make s2ram check
if there are active outputs on card0 in /sys/class/drm and skip the
quirks if that is the case.  Make s2both do an analogous thing.

Signed-off-by: Rafael J. Wysocki <rjw@sisk.pl>