s2disk: Do not fail hibernation withoug a VT console
Currently, s2disk fails to hibernate if there is no VT console
available to it, because it can't switch to a different VT in that
case.  However, switching VTs is not really essential and it should
be safe to carry out hibernation using the current console for
diagnostics.  For this reason, modify s2disk to prevent it from
failing when it can't use a VT console.

Signed-off-by: Rafael J. Wysocki <rjw@sisk.pl>