s2ram: Support quirks from configuration file
This patch permits suspend to RAM using quirks form the configuration
file. Options are:

s2ram_force=y	Force the suspend.
s2ram_quirks="quirks_options" Set the quirk options to suspend.

If the quirk is set in the configuration file, then the database is not
used.

To avoid the KMS suspension, needed to read these options from the
configuration file, use the "-k" parameter in s2ram. This parameter
can be used without options in the configuration file, to use the
machines in the database.

Other minor changes, like spaces at the end of the line are included too.

Signed-off-by: Rafael J. Wysocki <rjw@sisk.pl>